Proof of Concept
The combination and menu system is working as intended. Able to sort and separate the different elements into useful categories e.g. All, Combinable, and Creatable. Need to tweak the "minecraft" portion of the game so that the blocks are more easily placeable in the world. However, the way it is set up currently makes it easy to add and modify elements in both the combining menu and world space.